+++ src/main/java/com/qkm/TTMS/entity/MovieDirector.java
@@ -1,16 +1,12 @@
 package com.qkm.TTMS.entity;
 
-import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import java.io.Serializable;
 import lombok.Data;
 
-/**
- * 
- * @TableName movie_director
- */
+
 @TableName(value ="movie_director")
 @Data
 public class MovieDirector implements Serializable {
@@ -18,7 +14,7 @@ public class MovieDirector implements Serializable {
      * 主键
      */
     @TableId(value = "id")
-    private Long id;
+    private Integer id;
 
     /**
      * 导演的名字
@@ -36,7 +32,7 @@ public class MovieDirector implements Serializable {
      * 电影的id
      */
     @TableField(value = "movie_id")
-    private Long movieId;
+    private Integer movieId;
 
     @TableField(exist = false)
     private static final long serialVersionUID = 1L;
