+++ src/main/java/com/qkm/TTMS/entity/MovieUserRoles.java
@@ -1,16 +1,12 @@
 package com.qkm.TTMS.entity;
 
-import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import java.io.Serializable;
 import lombok.Data;
 
-/**
- * 
- * @TableName movie_user_roles
- */
+
 @TableName(value ="movie_user_roles")
 @Data
 public class MovieUserRoles implements Serializable {
@@ -18,13 +14,13 @@ public class MovieUserRoles implements Serializable {
      * 
      */
     @TableId(value = "id")
-    private Long id;
+    private Integer id;
 
     /**
      * 
      */
     @TableField(value = "user_id")
-    private Long userId;
+    private Integer userId;
 
     /**
      *
