+++ src/main/java/com/qkm/TTMS/entity/MovieWriter.java
@@ -1,16 +1,12 @@
 package com.qkm.TTMS.entity;
 
-import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import java.io.Serializable;
 import lombok.Data;
 
-/**
- * 
- * @TableName movie_writer
- */
+
 @TableName(value ="movie_writer")
 @Data
 public class MovieWriter implements Serializable {
@@ -18,13 +14,13 @@ public class MovieWriter implements Serializable {
      * 主键
      */
     @TableId(value = "id")
-    private Long id;
+    private Integer id;
 
     /**
      * 电影的id
      */
     @TableField(value = "movie_id")
-    private Long movieId;
+    private Integer movieId;
 
     /**
      * 编剧的照片
