+++ src/main/java/com/qkm/TTMS/entity/PeopleWant.java
@@ -1,16 +1,11 @@
 package com.qkm.TTMS.entity;
 
-import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableField;
-import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import java.io.Serializable;
 import lombok.Data;
 
-/**
- * 
- * @TableName people_want
- */
+
 @TableName(value ="people_want")
 @Data
 public class PeopleWant implements Serializable {
@@ -18,13 +13,13 @@ public class PeopleWant implements Serializable {
      * 主键id
      */
     @TableField(value = "id")
-    private Long id;
+    private Integer id;
 
     /**
      * 电影的id
      */
     @TableField(value = "movie_id")
-    private Long movieId;
+    private Integer movieId;
 
     /**
      * user id
