+++ src/main/java/com/qkm/TTMS/mapper/AreaCinemasMapper.java
@@ -2,26 +2,21 @@ package com.qkm.TTMS.mapper;
 
 import com.qkm.TTMS.entity.AreaCinemas;
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import com.qkm.TTMS.entity.CinemaMovies;
 import org.apache.ibatis.annotations.Param;
-import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Repository;
 
-import java.io.Serializable;
 import java.util.List;
 
-/**
- * @Entity com.qkm.TTMS.entity.AreaCinemas
- */
+
 @Repository
 public interface AreaCinemasMapper extends BaseMapper<AreaCinemas> {
-    List<AreaCinemas> getCinemaMoviesByCinemaId(@Param("areaName") String areaName, @Param("movieId") Long movieId);
-    AreaCinemas getAllById(@Param("id") Long id);
-    List<Long> getListID();
+    List<AreaCinemas> getCinemaMoviesByCinemaId(@Param("areaName") String areaName, @Param("movieId") int movieId);
+    AreaCinemas getAllById(@Param("id") int id);
+    List<Integer> getListID();
     List<AreaCinemas> getAll();
-    int deleteById(@Param("cinemaId") Long cinemaId);
-    int addMoney(@Param("money")Long money,@Param("cinemaId") Long cinemaId);
-    int downMoney(@Param("money")Long money,@Param("cinemaId") Long cinemaId);
+    int deleteById(@Param("cinemaId") int cinemaId);
+    int addMoney(@Param("money")Double money,@Param("cinemaId") int cinemaId);
+    int downMoney(@Param("money")Double money,@Param("cinemaId") int cinemaId);
     List<AreaCinemas> getAllByAreaName(@Param("cinemaName") String cinemaName);
 }
