+++ src/main/java/com/qkm/TTMS/mapper/MovieDirectorMapper.java
@@ -5,12 +5,10 @@ import com.qkm.TTMS.entity.MovieDirector;
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import org.springframework.stereotype.Repository;
 
-/**
- * @Entity com.qkm.TTMS.entity.MovieDirector
- */
+
 @Repository
 public interface MovieDirectorMapper extends BaseMapper<MovieDirector> {
-    int deleteByMovieId(@Param("movieId") Long movieId);
+    int deleteByMovieId(@Param("movieId") int movieId);
 }
