+++ src/main/java/com/qkm/TTMS/mapper/MovieUserMapper.java
@@ -6,16 +6,14 @@ import com.qkm.TTMS.entity.MovieUser;
 import org.apache.ibatis.annotations.Param;
 import org.springframework.stereotype.Repository;
 
-/**
- * @Entity com.qkm.TTMS.entity.MovieUser
- */
+
 @Repository
 public interface MovieUserMapper extends BaseMapper<MovieUser> {
     MovieUser getByAccounts(@Param("accounts") String accounts);
     MovieUser getAllByAccounts(@Param("accounts") String accounts);
    List<MovieUser> getAdminByCinemaId();
-   List<MovieUser> getSellByCinemaId(@Param("cinemaId") Long cinemaId);
-    int delById(@Param("id") Long id);
+   List<MovieUser> getSellByCinemaId(@Param("cinemaId") int cinemaId);
+    int delById(@Param("id") int id);
 }
