+++ src/main/java/com/qkm/TTMS/mapper/MovieUserRolesMapper.java
@@ -5,12 +5,10 @@ import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import org.apache.ibatis.annotations.Param;
 import org.springframework.stereotype.Repository;
 
-/**
- * @Entity com.qkm.TTMS.entity.MovieUserRoles
- */
+
 @Repository(value = "movieUserRolesMapper")
 public interface MovieUserRolesMapper extends BaseMapper<MovieUserRoles> {
-    int delByUserId(@Param("userId")Long userId);
+    int delByUserId(@Param("userId")int userId);
 }
