+++ src/main/java/com/qkm/TTMS/mapper/MovieWriterMapper.java
@@ -5,12 +5,10 @@ import com.qkm.TTMS.entity.MovieWriter;
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import org.springframework.stereotype.Repository;
 
-/**
- * @Entity com.qkm.TTMS.entity.MovieWriter
- */
+
 @Repository
 public interface MovieWriterMapper extends BaseMapper<MovieWriter> {
-    int deleteByMovieId(@Param("movieId") Long movieId);
+    int deleteByMovieId(@Param("movieId") int movieId);
 }
