+++ src/main/java/com/qkm/TTMS/mapper/UserOrderMapper.java
@@ -4,23 +4,20 @@ import java.util.List;
 import com.qkm.TTMS.entity.UserOrder;
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import org.apache.ibatis.annotations.Param;
-import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Repository;
 
-/**
- * @Entity com.qkm.TTMS.entity.UserOrder
- */
+
 @Repository
 public interface UserOrderMapper extends BaseMapper<UserOrder> {
-    List<UserOrder> getAllByCinemaId(@Param("cinemaId") Long cinemaId);
+    List<UserOrder> getAllByCinemaId(@Param("cinemaId") int cinemaId);
 
-    List<UserOrder> getAllByUserId(@Param("userId") Long userId);
+    List<UserOrder> getAllByUserId(@Param("userId") int userId);
 
-    int delById(@Param("id") Long id);
+    int delById(@Param("id") int id);
 
-    int updateOrderStatusById(@Param("orderStatus") String orderStatus, @Param("id") Long id);
+    int updateOrderStatusById(@Param("orderStatus") String orderStatus, @Param("id") int id);
 
-    int deleteByCinemaId(@Param("cinemaId") Long cinemaId);
+    int deleteByCinemaId(@Param("cinemaId") int cinemaId);
 
     int insertAll(UserOrder userOrder);
 }
