+++ src/main/java/com/qkm/TTMS/service/AreaCinemaService.java
@@ -7,13 +7,13 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 public interface AreaCinemaService {
-    List<AreaCinemas> getCinemaMoviesByCinemaId( String areaName, Long movieId);
-    AreaCinemas getAllById(Long id);
-    List<Long> getListID();
+    List<AreaCinemas> getCinemaMoviesByCinemaId( String areaName, int movieId);
+    AreaCinemas getAllById(int id);
+    List<Integer> getListID();
     List<AreaCinemas> getAll();
-    int deleteById(Long cinemaId);
-    int addMoney(Long money, Long cinemaId);
-    int downMoney(Long money, Long cinemaId);
+    int deleteById(int cinemaId);
+    int addMoney(Double money, int cinemaId);
+    int downMoney(Double money, int cinemaId);
     List<AreaCinemas> getAllByAreaName( String cinemaName);
 
 }
