+++ src/main/java/com/qkm/TTMS/service/CinemaMoviesService.java
@@ -1,16 +1,15 @@
 package com.qkm.TTMS.service;
 
 import com.qkm.TTMS.entity.CinemaMovies;
-import org.apache.ibatis.annotations.Param;
 
 import java.util.List;
 
 public interface CinemaMoviesService {
-    long getIdByCinemaIdAndMovieId(Long cinemaId,Long movieId);
-    int deleteByCinemaIdAndMovieId(Long cinemaId,Long movieId);
-    int  deleteByMovieId(Long movieId);
-    List<Long> getListMovieIdByCinemaId(Long cinemaId);
-    int setMovieLowMoneyByCinemaIdAndMovieId(Double movieLowMoney,Long cinemaId,Long movieId);
-    CinemaMovies getAllByCinemaId( Long cinemaId);
-    int deleteByCinemaId( Long cinemaId);
+    int getIdByCinemaIdAndMovieId(int cinemaId,int movieId);
+    int deleteByCinemaIdAndMovieId(int cinemaId, int movieId);
+    int  deleteByMovieId(int movieId);
+    List<Integer> getListMovieIdByCinemaId(int cinemaId);
+    int setMovieLowMoneyByCinemaIdAndMovieId(Double movieLowMoney,int cinemaId,int movieId);
+    CinemaMovies getAllByCinemaId( int cinemaId);
+    int deleteByCinemaId( int cinemaId);
 }
