+++ src/main/java/com/qkm/TTMS/service/MoviePlanService.java
@@ -6,5 +6,5 @@ import org.apache.ibatis.annotations.Param;
 import java.util.List;
 
 public interface MoviePlanService {
-    List<MoviePlan> getMoviePlan(Long movieId,Long cinemaId);
+    List<MoviePlan> getMoviePlan(int movieId,int cinemaId);
 }
